+++ toolkitplugin/src/main/java/org/fourz/rvnktools/command/manager/BaseSubCommand.java
@@ -1,4 +1,4 @@
-package org.fourz.rvnktools.command.framework;
+package org.fourz.rvnktools.command.manager;
 
 import org.bukkit.ChatColor;
 import org.bukkit.command.CommandSender;
