+++ toolkitplugin/src/main/java/org/fourz/rvnktools/command/manager/BroadcastCommand.java
@@ -1,10 +1,9 @@
-package org.fourz.rvnktools.command.framework.commands;
+package org.fourz.rvnktools.command.manager;
 
 import org.bukkit.Bukkit;
 import org.bukkit.ChatColor;
 import org.bukkit.command.CommandSender;
 import org.fourz.rvnktools.RVNKTools;
-import org.fourz.rvnktools.command.framework.BaseCommand;
 
 import java.util.Collections;
 import java.util.List;
