+++ toolkitplugin/src/main/java/org/fourz/rvnktools/command/manager/CommandValidator.java
@@ -1,4 +1,4 @@
-package org.fourz.rvnktools.command.framework;
+package org.fourz.rvnktools.command.manager;
 
 import org.bukkit.command.CommandSender;
 import org.bukkit.entity.Player;
