+++ toolkitplugin/src/main/java/org/fourz/rvnktools/command/manager/TrainsCommand.java
@@ -1,9 +1,9 @@
-package org.fourz.rvnktools.command.framework.commands;
+package org.fourz.rvnktools.command.manager;
 
 import org.bukkit.command.CommandSender;
 import org.bukkit.entity.Player;
 import org.fourz.rvnktools.RVNKTools;
-import org.fourz.rvnktools.command.framework.BaseCommand;
+
 import java.util.Collections;
 import java.util.List;
 import java.util.Arrays;
